refactor: split mesh into smaller bits
